fixed dtype issue

+++ recursive_lm/optimizer.py
@@ -49,7 +49,7 @@ def normuon_update(grad, momentum, second_momentum, beta=0.95, beta2=0.95, ns_st
         update = update.reshape(original_shape)
     ################ NorMuon added ###################
     vnorm = update.norm(dim=(-2,-1), keepdim=True)
-    v_mean = torch.mean(update * update, dim=-1, keepdim=True)
+    v_mean = torch.mean(update * update, dim=-1, keepdim=True, dtype=second_momentum.dtype)
     second_momentum.lerp_(v_mean, 1 - beta2)
     step_size = 1 / second_momentum.sqrt().add_(1e-10)
     update.mul_(step_size)
